Improve Pattern query config generation
